Can now compile myopic world

+++ refresh.sh
@@ -6,11 +6,14 @@ AIPATH=~/Libraries/Self/AIToolbox
 
 if [ $# -ge 1 ]
 then
-    ( cd "$AIPATH"/build && make -j )
-    rm -rf include
-    cp -ru "$AIPATH"/include .
-    rm ./lib/*
-    cp -u "$AIPATH"/build/lib* ./lib/
+    if [ $# -ge 2 ]
+    then
+        ( cd "$AIPATH"/build && make -j )
+        rm -rf include/AIToolbox
+        cp -ru "$AIPATH"/include .
+        rm ./lib/*
+        cp -u "$AIPATH"/build/lib* ./lib/
+    fi
 
     cd build
     rm -r *
